refatorando2
