+++ include/CMakeLists.txt
@@ -1,3 +1,5 @@
 target_sources(${PROJECT_NAME} PRIVATE
     Color.hpp
+    Ray.hpp
+    Vector3d.hpp
 )
